backend/requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
Flask-PyMongo
Flask-CORS
pandas
numpy
torch
torchvision
matplotlib
seaborn
docplex
pyomo
scipy
pymongo
ipykernel
python-dotenv
Werkzeug~=2.0.0
geopy
bcrypt
googlemaps
